Further code splitting and refactoring, putting the jar entry extraction in a different class.
